requirements.txt
flask==1.1.2
Flask-Cors==3.0.9
python-dotenv==0.14.0
Pyrebase==3.0.27
urllib3==1.25.11
requests==2.24.0
firebase-admin==4.4.0
Cerberus==1.3.2
flask-swagger-ui==3.25.0
gunicorn==20.0.4